Move SkinnedMesh to mesh
