+++ pybevy/animation.pyi
@@ -371,53 +371,6 @@ class AnimationTarget:
     id: AnimationTargetId
     player: Entity
     def __init__(self, id: AnimationTargetId, player: Entity) -> None: ...
-
-class SkinnedMesh(Component):
-    """Component that defines a skinned mesh for skeletal animation.
-
-    A skinned mesh deforms its vertices based on a hierarchy of bone entities (joints).
-    Each vertex is influenced by one or more bones, weighted by the inverse bind pose matrices.
-
-    This component is typically created automatically when loading GLTF files with skeletal
-    animations, but can also be constructed manually for procedural skinned meshes.
-    """
-
-    def __init__(self, inverse_bindposes: Handle, joints: list[Entity]) -> None:
-        """Create a new SkinnedMesh component.
-
-        Args:
-            inverse_bindposes: Handle to SkinnedMeshInverseBindposes asset containing
-                the inverse bind pose matrices for each joint
-            joints: List of Entity references representing the bone hierarchy
-        """
-
-    @property
-    def inverse_bindposes(self) -> Handle:
-        """Get the handle to the inverse bind pose matrices asset."""
-
-    @property
-    def joints(self) -> list[Entity]:
-        """Get the list of joint (bone) entities."""
-
-    def joint_count(self) -> int:
-        """Get the number of joints in the skeleton."""
-
-    def get_joint(self, index: int) -> Entity:
-        """Get a specific joint entity by index.
-
-        Args:
-            index: Zero-based index of the joint to retrieve
-
-        Returns:
-            The Entity at the specified index
-
-        Raises:
-            ValueError: If index is out of bounds
-        """
-
-    def __len__(self) -> int:
-        """Get the number of joints in the skeleton."""
-
 class AnimatedBy(Component):
     """Component indicating which AnimationPlayer entity drives this entity's animations.
